refactor(daco-26): revert App.tsx name and change HomePage.tsx to index.tsx
- change index.tsx to App.tsx
- change HomePage.tsx to index.tsx
